Fix an error when open external music

+++ app/src/main/java/com/naman14/timber/MusicPlayer.java
@@ -658,7 +658,8 @@ public class MusicPlayer {
 
     public static void clearQueue() {
         try {
-            mService.removeTracks(0, Integer.MAX_VALUE);
+            if (mService != null)
+                mService.removeTracks(0, Integer.MAX_VALUE);
         } catch (final RemoteException ignored) {
         }
     }
